+++ packages/tests/package.json
@@ -13,6 +13,9 @@
         "import": "./dist/index.js"
       }
     },
+    "dependencies": {
+      "csp-toolkit": "workspace:^"
+    },
     "devDependencies": {
         "@rollup/plugin-node-resolve": "^15.2.3",
         "@rollup/plugin-typescript": "^11.1.6",
