+++ packages/tests/tsconfig.json
@@ -1,4 +1,18 @@
 {
-  "extends": "@repo/typescript-config/base",
-  "include": ["src"],
+  "extends": "@repo/typescript-config/base.json",
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "module": "ESNext",
+    "target": "esnext",
+    "moduleResolution": "Bundler",
+    "strict": true,
+    "noUnusedLocals": true,
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
+    "noImplicitAny": true,
+    "lib": ["dom", "esnext"],
+    "types": ["node"],
+    "noEmitOnError": true,
+  },
+  "include": ["src/**/*"],
 }
