+++ packages/vite-plugin-csp-guard/rollup.config.mjs
@@ -18,7 +18,7 @@ export default defineConfig({
       sourcemap: false,
     },
   ],
-  external: ["lightningcss", "fsevents", "vite", "cheerio"],
+  external: ["lightningcss", "fsevents", "vite", "cheerio", "csp-toolkit"],
   plugins: [
     typescript({
       tsconfig: "tsconfig.json",
